[Chore] #38 LoginView 로직 관련 코드 LoginStateModel로 이동
