+++ IAteIt/IAteIt/View/SignUp/Model/LoginStateModel.swift
@@ -5,6 +5,8 @@
 //  Created by Eunbee Kang on 2023/04/26.
 //
 
+import AuthenticationServices
+import CryptoKit
 import Foundation
 import Firebase
 import FirebaseAuth
@@ -15,6 +17,7 @@ class LoginStateModel: ObservableObject {
     @Published var user: User?
     @Published var isAppleLoginRequired: Bool = false
     @Published var isSignUpViewPresent: Bool = false
+    @Published var isSignUpRequired: Bool = false
     
     init() {
         self.checkLoginUser()
@@ -36,6 +39,59 @@ class LoginStateModel: ObservableObject {
             }
         } else {
             self.isAppleLoginRequired = true
+        }
+    }
+    
+    func getResultOfAppleLogin(result: Result<ASAuthorization, Error>, currentNonce: String?) {
+        switch result {
+        case .success(let authResults):
+            print("Authorisation successful \(authResults)")
+          
+            switch authResults.credential {
+            case let appleIDCredential as ASAuthorizationAppleIDCredential:
+                guard let nonce = currentNonce else {
+                    fatalError("Invalid state: A login callback was received, but no login request was sent.")
+                }
+                guard let appleIDToken = appleIDCredential.identityToken else {
+                    fatalError("Invalid state: A login callback was received, but no login request was sent.")
+                }
+                guard let idTokenString = String(data: appleIDToken, encoding: .utf8) else {
+                    print("Unable to serialize token string from data: \(appleIDToken.debugDescription)")
+                    return
+                }
+                let credential = OAuthProvider.credential(withProviderID: "apple.com", idToken: idTokenString, rawNonce: nonce)
+                Auth.auth().signIn(with: credential) { (authResult, error) in
+                    if error != nil {
+                        // Error. If error.code == .MissingOrInvalidNonce, make sure
+                        // you're sending the SHA256-hashed nonce as a hex string with
+                        // your request to Apple.
+                        print(error?.localizedDescription as Any)
+                        return
+                    }
+                    if let authUser = authResult?.user {
+                        print("애플 로그인 결과:", authUser.uid, authUser.email ?? "-")
+                        FirebaseConnector().checkExistingUser(userUid: authUser.uid) { isExist in
+                            if isExist {
+                                FirebaseConnector().fetchUser(id: authUser.uid) { fetchedUser in
+                                    self.user = fetchedUser
+                                    self.isAppleLoginRequired = false
+                                }
+                            } else {
+                                self.appleUid = authUser.uid
+                                self.isAppleLoginRequired = false
+                                self.isSignUpRequired = true
+                            }
+                        }
+                    }
+                }
+                print("\(String(describing: Auth.auth().currentUser?.uid))")
+                
+            default:
+                break
+            }
+            
+        case .failure(let error):
+          print("Authorisation failed: \(error.localizedDescription)")
         }
     }
 }
